Update patch set 1
Patch Set 1:

> @jpena, I think the same type of changes required for all projects?

Actually, we'll probably want to give it a deeper look. oslo-reports is built in Fedora too, and we don't want to use openstack-macros in packages built in Fedora to minimize differences.

For this specific case, openstack-macros was only listed as a python3 subpackage BR, which should not be a common case.

We could merge this as-is, and then start an easyfix project to scan all packages built in Fedora to remove openstack-macros usage.

Patch-set: 1